+++ samples/smallluxgpu4/smallluxgpu.cpp
@@ -139,7 +139,7 @@ static int BatchSimpleMode(const double haltTime, const u_int haltSpp, const flo
 		LC_LOG(boost::str(boost::format("[Elapsed time: %3d/%dsec][Samples %4d/%d][Convergence %f%%][Avg. samples/sec % 3.2fM on %.1fK tris]") %
 				int(elapsedTime) % int(haltTime) % pass % haltSpp % (100.f * convergence) %
 				(stats.Get("stats.renderengine.total.samplesec").Get<double>() / 1000000.0) %
-				(stats.Get("stats.dataset.trianglecount").Get<size_t>() / 1000.0)));
+				(stats.Get("stats.dataset.trianglecount").Get<double>() / 1000.0)));
 
 	}
